refactor(pluggable-widgets-tools): move typing generation to rollup
