+++ packages/tools/pluggable-widgets-tools/configs/shared.js
@@ -1,0 +1,10 @@
+import { promises as fs } from "fs";
+import { join } from "path";
+
+export async function listDir(path) {
+    const entries = await fs.readdir(path, { withFileTypes: true });
+    return entries
+        .filter(e => e.isFile())
+        .map(e => join(path, e.name))
+        .concat(...(await Promise.all(entries.filter(e => e.isDirectory()).map(e => listDir(join(path, e.name))))));
+}
