+++ packages/tools/pluggable-widgets-tools/package.json
@@ -77,7 +77,6 @@
     "jasmine-core": "^3.6.0",
     "jest": "^26.6.1",
     "jest-junit": "^12.0.0",
-    "map-stream": "^0.0.7",
     "mendix": "^8.15.7685",
     "metro-react-native-babel-preset": "~0.63.0",
     "node-fetch": "^2.6.1",
@@ -103,8 +102,6 @@
     "yazl": "^2.5.1"
   },
   "devDependencies": {
-    "@types/plugin-error": "^0.1.1",
-    "@types/vinyl": "^2.0.4",
     "@types/xml2js": "^0.4.5",
     "@mendix/generator-widget": ">=8.9.0",
     "async-mutex": "^0.2.4",
