--- packages/tools/pluggable-widgets-tools/src/typings-generator/map-stream.d.ts
@@ -1,5 +1,0 @@
-declare module "map-stream" {
-    function map(dataFunction: (data: any, callback: () => void) => void): any;
-
-    export = map;
-}
